fix schedule task
